PEP8 #100 Rigbits: refactor finish
